test(order): 주문·재고·쿠폰 동시성 통합 스펙(여러명/동일사용자 × 성공/실패) 추가

+++ apps/commerce-api/src/main/kotlin/com/loopers/infrastructure/like/LikeRepositoryImpl.kt
@@ -21,7 +21,7 @@ class LikeRepositoryImpl(
         likeJpaRepository.existsByUserIdAndProductId(userId, productId)
 
     override fun delete(like: Like): Long =
-        likeJpaRepository.deleteByUserIdAndProductId(like.userId, like.productId)
+        likeJpaRepository.deleteByUserIdAndProductId(like.userId, like.productId).toLong()
 
     override fun findAllByUserId(userId: Long, page: Int, size: Int): PageResult<Like> {
         val found = likeJpaRepository.findAllByUserId(
